projects/policyengine-api-simulation/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "policyengine-simulation-api-project"
version = "0.1.0"
readme = "README.md"
authors = [
    {name = "PolicyEngine", email = "hello@policyengine.org"},
]
license = {file = "../../LICENSE"}
requires-python = ">=3.13,<3.14"
dependencies = [
    "opentelemetry-instrumentation-sqlalchemy (>=0.51b0,<0.52)",
    "pydantic-settings (>=2.7.1,<3.0.0)",
    "opentelemetry-instrumentation-fastapi (>=0.51b0,<0.52)",
    "policyengine-fastapi",
    "policyengine==0.8.1",
    "policyengine-uk>=2.22.8",
    "policyengine-us>=1.370.2",
    "tables>=3.10.2",
    "modal>=0.73.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/policyengine_api_simulation", "src/policyengine_api"]

[tool.uv.sources]
policyengine-fastapi = { path = "../../libs/policyengine-fastapi", editable = true }

[project.optional-dependencies]
test = [ "pytest>=8.3.4", "pytest-asyncio>=0.25.3", "pytest-cov>=6.1.1",]
build = [ "pyright>=1.1.401", "black>=25.1.0", "openapi-python-client>=0.21.6",]

[tool.pytest.ini_options]
pythonpath = [
  "src",
]

testpaths = ["tests"]
